refactor: move utility functions to common/utils and update imports
