+++ src/common/utils/string-utils.ts
