+++ src/schemas/email-schemas.ts
@@ -1,6 +1,6 @@
 import { MsgType } from "./msg-type";
 import { z } from "zod";
-import { trimOrUndefined } from "../utils/string-utils";
+import { trimOrUndefined } from "../common/utils/string-utils";
 import { formatErrorMessage } from "./message-handler";
 import { EMAIL_PATTERN } from "../common/regex-patterns";
